requirements.txt
anthropic>=0.40.0
youtube-transcript-api>=0.6.2
yt-dlp>=2024.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
requests>=2.31.0
edge-tts>=7.0.0
playwright>=1.40.0
